Adjust the API to provide more feedback.
Restore the concept of the Success/Failure message after conversion
being the response to the Set Filename request. If successful, Set
Filename will no longer send a Success response.

Add an error code to the Failure response, to indicate the cause of
the failure.

Test for NULL filenames, and reject them.

Update the documentation to include details of the API.